Modernize buildscript
Used modern Gradle features such as the incubating JVM Testing Suite
plugin to streamline declarations. The subproject configurations are
still contained inside of the main `build.gradle` and are written in a
way that doesn't cause configuration cache issues.

Since we plan on rewriting Jar-in-Jar in the (near) future, I'm not
worried about making this project compliant with isolated projects. So
using `Project#subprojects` in the main buildscript is fine.

Other changes
- Enabled configuration cache
- `src/binks` -> `src/test`
- Test failures are now ignored.

Dependency changes:
- Gradle 9.0.0
- GradleUtils 3.3.0
- Git Version Gradle 3.1.0
- Changelog Gradle 3.1.1