gradle-0.1 - Gradle plugin (#7)
Forge Jar-in-Jar Gradle is a standalone Gradle plugin that enables
buildscripts to embed dependencies using Forge's Jar-in-Jar
specification with minimal hassle.

While primarily designed for Forge mods and only uses Forge's Jar-in-Jar
specification, this can be used in any Gradle project without being
strictly tied to ForgeGradle, allowing for complex setups that are not
reliant on the presence of ForgeGradle.

Despite being functional, this is still **work in progress!** Expect
bugs, and for things to might not work. Documentation will come at a
later date. For examples on how to use, see `jarjar-gradle-demo` in the
root directory.